move ApiEndpoint stuff to dir, ApiCall gets own file, types for all
